+++ pkg/upload/dart.go
@@ -19,7 +19,7 @@ import (
 var androidSymbolFileRegex = regexp.MustCompile("android-([^;]*).symbols")
 var iosSymbolFileRegex = regexp.MustCompile("ios-([^;]*).symbols")
 
-func Dart(options options.CLI, endpoint string, logger log.Logger) error {
+func Dart(options options.CLI, logger log.Logger) error {
 	dartOptions := options.Upload.DartSymbol
 	fileList, err := utils.BuildFileList(dartOptions.Path)
 
@@ -44,12 +44,12 @@ func Dart(options options.CLI, endpoint string, logger log.Logger) error {
 			}
 
 			// Build Upload options
-			uploadOptions := utils.BuildDartUploadOptions(options.ApiKey, buildId, "android", options.Upload.Overwrite, dartOptions.VersionName, dartOptions.VersionCode)
+			uploadOptions := utils.BuildDartUploadOptions(buildId, "android", options.Upload.Overwrite, dartOptions.VersionName, dartOptions.VersionCode)
 
 			fileFieldData := make(map[string]server.FileField)
 			fileFieldData["symbolFile"] = server.LocalFile(file)
 
-			err := server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, file, options, logger)
+			err := server.ProcessFileRequest(options.ApiKey, "/dart-symbol", uploadOptions, fileFieldData, file, options, logger)
 
 			if err != nil {
 
@@ -85,7 +85,7 @@ func Dart(options options.CLI, endpoint string, logger log.Logger) error {
 			}
 
 			// Build Upload options
-			uploadOptions := utils.BuildDartUploadOptions(options.ApiKey, buildId, "ios", options.Upload.Overwrite, dartOptions.VersionName, dartOptions.BundleVersion)
+			uploadOptions := utils.BuildDartUploadOptions(buildId, "ios", options.Upload.Overwrite, dartOptions.VersionName, dartOptions.BundleVersion)
 
 			fileFieldData := make(map[string]server.FileField)
 			fileFieldData["symbolFile"] = server.LocalFile(file)
@@ -93,7 +93,7 @@ func Dart(options options.CLI, endpoint string, logger log.Logger) error {
 			if options.DryRun {
 				err = nil
 			} else {
-				err = server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, file, options, logger)
+				err = server.ProcessFileRequest(options.ApiKey, "/dart-symbol", uploadOptions, fileFieldData, file, options, logger)
 			}
 
 			if err != nil {